Add markdown + frontmatter support
Adds .md files for index & about routes.
