+++ app/about/about.md
@@ -1,0 +1,16 @@
+---
+title: About
+---
+
+This is the home of computing and software at ISIS Neutron and Muon Facility.
+
+You can find out more info about ISIS, as well as scientific documentation at [isis.stfc.ac.uk](https://www.isis.stfc.ac.uk/)
+
+You can find the source code for various projects at GitHub:
+[isisneutronmuon](https://github.com/isisneutronmuon)
+[isisbusapps](https://github.com/isisbusapps/)
+[ral-facilities](https://github.com/ral-facilities)
+
+You can find the source code for Jekyll and the Minima theme used here at GitHub:
+[jekyll](https://github.com/jekyll/jekyll) /
+[minima](https://github.com/jekyll/minima)
