+++ app/index.md
@@ -1,0 +1,22 @@
+Software that enables researchers to process and interpret the experimental data from more than 30 different instruments at the ISIS facility. This software is essential for advancing scientific research and innovation in various fields and domains.
+
+* [Aten](https://github.com/trisyoungs/aten) for editing atomic configurations such as those used by, and generated from, [EPSR](https://www.isis.stfc.ac.uk/Pages/Empirical-Potential-Structure-Refinement.aspx)
+* [Dissolve](https://projectdissolve.com/) for the interrogation of scattering data, typically total neutron scattering data
+* [dlputils](https://github.com/trisyoungs/dlputils) for calculation of quantities from, and manipulation of, DL_POLY/xyz trajectories
+* [Euphonic](https://github.com/pace-neutrons/Euphonic) for efficient simulation of phonon bandstructures, density of states and inelastic neutron scattering intensities from force constants.
+* [Fitbenchmarking](https://fitbenchmarking.github.io/) for comparing different minimizers/fitting frameworks.
+* [GudPy](https://github.com/disorderedmaterials/GudPy) Python-based batch processing and alternative GUI for Gudrun also offering additional functionality unavailable in the Gudrun Java GUI
+* [Gudrun](https://github.com/disorderedmaterials/Gudrun) for reducing data from neutron and x-ray scattering experiments
+* [HOGBEN](https://github.com/jfkcooper/HOGBEN) for optimising the design of a wide range of reflectometry experiments.
+* [Horace](https://pace-neutrons.github.io/Horace/stable/) for the visualisation and analysis of large datasets from time-of-flight neutron inelastic scattering spectrometers.
+* [Mantid](http://www.mantidproject.org/) for processing, particularly reducing, materials-science data.
+* [McPhase](http://mcphase.de/) for the calculation of magnetic properties.
+* [MDANSE](https://www.isis.stfc.ac.uk/Pages/MDANSEproject.aspx) for computing neutron observables from molecular dynamics (MD) trajectories that can be directly compared with neutron scattering experiments, particularly inelastic and quasi-elastic neutron scattering spectroscopies.
+* [MDMC](https://github.com/MDMCproject) for automating the typical workflow of researchers that create atomistic/microscopic models (via molecular dynamics, MD, simulations) and compare them with experimental data (mostly from neutrons scattering especially QENS).
+* [MuSSIC](https://github.com/disorderedmaterials/MuSSIC) for calculating neutron-weighted structure factors from coarse-grained simulation trajectories.
+* [PACE](https://www.isis.stfc.ac.uk/Pages/Proper-analysis-of-coherent-excitations.aspx) in development and will provide an integrated visualisation, simulation and fitting environment on massively parallel and distributed computing, interfacing to third party modelling codes.
+* [RAMP](https://github.com/gcassella/RAMP) a Monte Carlo raytracing package for the simulation of neutron instrumentation.
+* [RasCAL](https://github.com/arwelHughes/RasCAL_2019) for analysing data collected on ISIS reflectometry beamlines.
+* [SScanSS 2](https://github.com/ISISNeutronMuon/SScanSS-2) for planning, visualising, and setting-up strain scanning experiments on engineering beam-line instruments.
+* [SasView](https://www.sasview.org/) for the analysis of small angle scattering data.
+* [SpinW](https://spinw.org/) to plot and numerically simulate magnetic structures and excitations of given spin Hamiltonian using classical Monte Carlo simulation and linear spin wave theory.
