--- next.config.js
@@ -1,9 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  basePath: '',
-  cleanDistDir: true,
-  distDir: './dist',
-  output: 'export'
-}
-
-module.exports = nextConfig
