+++ tsconfig.json
@@ -33,7 +33,8 @@
     "**/*.ts",
     "**/*.tsx",
     ".next/types/**/*.ts",
-    "./dist/types/**/*.ts"
+    "./dist/types/**/*.ts",
+    "app/page.mdx"
   ],
   "exclude": [
     "node_modules",
